requirements.txt
numpy
matplotlib
lmfit
scipy
pyyaml
